detection/requirements-gpu.txt
opencv-python==4.5.2.52
tensorflow-gpu==2.4.1
scikit-image